Release

+++ app/controllers/skills_controller.rb
@@ -30,6 +30,7 @@ class SkillsController < ApplicationController
   end
 
   def update
+    authorize @skill
     if @skill.update skill_params
       flash[:notice] = 'Successfully updated'
     else
@@ -39,12 +40,18 @@ class SkillsController < ApplicationController
   end
 
   def create
-    Skill.create skill_params
-    flash[:notice] = 'Successfully created'
+    @skill = Skill.new(skill_params)
+    authorize @skill
+    if @skill.save
+      flash[:notice] = 'Successfully created'
+    else
+      flash[:danger] = 'Is not created [duplicate]'
+    end
     redirect_to skills_path
   end
 
   def destroy
+    authorize @skill
     if @skill.destroy
       flash[:notice] = 'Successfully destroyed'
     else
@@ -61,5 +68,9 @@ class SkillsController < ApplicationController
 
   def skill_params
     params.require(:skill).permit!
+  end
+
+  def pundit_user
+    current_employee
   end
 end
